split instrumented test files
